Update patch set 34
Patch Set 34:

So, the plan David and I discussed was to target draft 4 (the latest) in this CL, then write a followup to bring it up to a future draft when it comes out.

It was a great plan, but then I helped Chris verify the new test vectors in https://github.com/cisco/go-hpke/pull/30. Out of laziness, I'd like to wait for a new draft to be cut (or at least for PR#30 to be merged).

Honestly, the hardest thing to get right between these revisions has been the test vector format!

Patch-set: 34