Update patch set 1
Patch Set 1: Commit-Queue+1

The concern here, and the reason that it's not enabled already, is that we worry about being able to test this code. I don't expect that any of the Chromium builders support SHAEXT. Wikipedia says that it's included in Ice Lake, however, and our SDE config includes that.

Let's see if the SDE builders are happy...

Patch-set: 1
Reviewer: Gerrit User 5005 <5005@66b711a4-aa9b-3dac-9062-916540830212>
Label: Commit-Queue=+1