Update patch set 1
Patch Set 1: Code-Review+2

Feck. I installed an up-to-date copy of libunwind and that fixed it.

With that, there's a silly CFI error in the assembly which I'll fix and upstream in another change. Otherwise, I think this is good to go.

David: ok with this? We should have testing via SDE and the unwind tests (will) pass.

Patch-set: 1
Reviewer: Gerrit User 5005 <5005@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=+2
Label: Commit-Queue=0