Support disabling -Werror during build
Depending on the configuration and compiler used while building
BoringSSL, spurious warnings might be generated. This can cause build
failures for projects that integrate BoringSSL, e.g.
https://github.com/cloudflare/quiche/issues/589

In these cases warnings are not useful as they can't really be actioned
but due to how CMake works adding "-Wno-error" when invoking CMake
doesn't work, so this change adds an option to avoid adding "-Werror" to
begin with.

The -Werror option is still enabled by default, so this shouldn't affect
development of BoringSSL.

Change-Id: Ie18a5070f6b31bef9dc664dec8df137b41c52fec