Handle NULL arguments in some i2d_* functions.
Prior to 5d7c2f8b1d, these i2d functions would fail, rather than crash,
if passed a NULL argument. While we don't generally have much truck with
the idea that functions should be expected to handle NULL arguments
where not documented, it seems that a fair amount of code is depending
on this.

Change-Id: I928b35533aa2a7beed57d7f58ba44681a8eb05c6