acvp: support working with files.
A direct connections to the ACVP servers may not always be available. In
some cases, NVLAP labs will interact with the servers and send JSON back
and forth as files. This change supports both dumping the capabilities
JSON (which a lab will need in order to send to the server) and
processing vectors from a file on disk.

Change-Id: Iefa0c411b9a19808b5a7eb431169068d1c2ea966