Const-correct X509V3_CONF_METHOD.
This is needed to fix all the config APIs to take const char *. I've
split it out as it's the only incompatible half of the change.

Update-Note: External definitions of X509V3_CONF_METHOD will need fix
the types of their functions. There should not be any of these (probably
hide this struct), but if there are, this aligns with upstream OpenSSL.

Change-Id: I6e760cfbca5d3f408215b8f3744acd1fd7f31391