Remove sk_new_null call.
This entire function is assuming all the STACK_OF(T) types are secretly
the same type, but best to consistently use the sk_ASN1_VALUE_*
wrappers. The raw sk_foo functions are an implementation detail of the
macros and we probably should rename them to be better prefixed (as
upstream did).

Change-Id: I62d910b93ca6be5e1c83ae269c7df6a437ffb316