Allow specifying different initial and resumption expectations.
We have an ad-hoc expectResumeVersion, and I'll want to add one for ALPS
(to confirm changing ALPS on resumption works). We probably could have
done with such a test for ALPN too. Wrap all the ConnectionState
expectations into a struct with an optional resumeExpectations field.

Also move the OCSPResponse() method to ConnectionState for consistency.

Change-Id: Icaabf5571c51e78ed078f57de0e04928d3f3fa8d