acvp: support saving vectors.
Since we have support for reading vectors from files, this change adds
support for saving them. There's no support for uploading the saved
vectors, rather it's just for quicker debugging since the NIST server is
taking over a minute to produce vectors at the moment and that's a
little frustrating to iterate with.

Change-Id: I5da8a084eb06b81aefa838b4e7ad8d529d1d31a6