runner: Implement a more complete ClientHello consistency check.
The Go TLS implementation, at the time runner forked, had custom
testing-only equal methods on all the handshake messages. We've since
removed all of them except for ClientHello, where we repurposed the
function to check ClientHello consistency on HelloVerifyRequest and
HelloRetryRequest.

These are tedious to update. Upstream has since replaced them with
reflect.DeepEqual, but the comparison we want is even tighter. Even
unknown extensions aren't allowed to change. Replace the check with a
custom one that works on the byte serialization and remove
clientHelloMsg.equal.

Along the way, I've fixed the HRR PSK identity logic to match the spec a
bit more and check binders more consistently.

Change-Id: Ib39e8791201c42d37e304ae5110c7aeed62c8b3f