emscripten cross-compilation support
This change simplifies embedding boringssl as a ssl/crypto library for
applications deployed as webassembly moduldes. The tests are exceptionally
large binaries that require fiddling with emscriptens memory configurations
defaults.

To embed the library with cmake you can now do the following:
 - disable assembly code
 - disable tests
 - define an alias OpenSSL::SSL
 - define target include paths
 - include the boringssl CMakeLists.txt

With that the library is more or less compatible to FindPackage(OpenSSL)
  if(EMSCRIPTEN)
     set(OPENSSL_NO_ASM "1")
  endif()

  set(BORINGSSL_DISABLE_TEST ON CACHE BOOL "Disable test building" FORCE)
  add_subdirectory(boringssl)
  target_compile_options(ssl PRIVATE -Wno-format)
  target_include_directories(ssl
      PUBLIC
          $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/boringssl/include/>
          $<INSTALL_INTERFACE:include/>
          )
  add_library(OpenSSL::SSL ALIAS ssl)

Change-Id: I6987129eeea2513e4c1c35f0e97f8c2ea60ec1b8