Fix crash when flushing an SSL BIO.
OpenSSL synchronizes bio->next_bio and ssl->rbio with a variety of
callbacks, so BIO_copy_next_retry worked. We do not, so attempting to
flush the BIO crashed.

The SSL BIO is a compatibility hack and intentionally much more limited,
so start by just copying things from the right BIO directly. Add a basic
unit test for SSL BIOs. If we need to, we can implement a more complex
synchronization later.

Additionally reject reconfiguring an SSL BIO because that will leak the
object right now.

Change-Id: I724c95ab6f1a3a1aa1889b0483c81ce3bdc534ae