Add SSL_early_data_reason_string.
QUICHE has a switch-case converting ssl_early_data_reason_t to a string
for logging. This causes a lot of churn when we add a new value.
Instead, add a function for this. Bump BORINGSSL_API_VERSION so we can
easily land a CL in QUICHE to start using the function without
coordinating repositories.

Change-Id: I176ca07b4f75a3ea7153a387219459665062aad9