Add util/fetch_echconfigs.go.
This tool queries DoH for a given domain's HTTPS records. It parses the
response and attempts to extract an ECHConfigs structure if one is
present.

Currently, this tool requests a JSON response. We assume the HTTPS
resource record will be represented as a hexadecimal string, but this is
pretty unstable and could change at the whim of the resolver.

Once golang.org/x/net/dns/dnsmessage supports parsing of unknown RR
types, we will be able to request the actual bytes of the DNS message
("Accept: application/dns-message").

Bug: 275
Change-Id: I705591658921f60a958164a18b68ffb697c2ea4b