acvp: update to newer FFDH test.
Revision 1.0 is this test is reportedly no longer acceptable and we have
to use the “SSC” version now. The documentation for this test doesn't
mention that a “z” field is possible, but that's what the test vectors
from the demo server contain and, after guessing at the correct response
format, this makes the NIST server happy.

Change-Id: Ic63d9e19998dc015733d847cd0330a3af1d5e7e6