acvp: move CMAC verification into the module wrapper.
CMAC is not inside our FIPS module and we have ACVP support for it just
for testing (other modules need to validate CMAC). This change makes the
CMAC verify test an explicit action for the module wrapper so that a
verification function exposed by a FIPS module can be tested.

Change-Id: I3943bde175f2c1d62881002b4e12d7bca68a9018