Fix TLS13SessionID-TLS13 test.
The check was happening in code that only ran at TLS 1.2, so we weren't
testing anything. Additionally check the resumption case. While we do
handle it correctly, we only manage it due to the weird OpenSSL quirk
we've carried over from TLS 1.2 tickets where we synthesize a session ID
for TLS 1.3 tickets. (Is that still needed?)

That's subtle enough to warrant a test, and some other implementations
reuse our test suite, so it's worth the coverage there.

Change-Id: I83cc355bd853097ec6edcd2cc40b06b19e3b00e5