Add ECH server (draft-ietf-tls-esni-09).
This CL adds a complete implementation of the ECH server, with pieces of
the BoGo client as necessary for testing. In particular, the server
supports ClientHelloInner compression with ech_outer_extensions. When
ECH decryption fails, it can send retry_configs back to the client.

This server passes a simple interop test against the cloudflare-go and
nss clients. See the "ech-accept" test case in tls-interop-runner [0].

This CL replaces two prior WIP CLs. In the first one, I was working on
adding ECH client and server behavior to BoGo [1]. In the second one, I
was adding just the C server [2].

[0]: https://github.com/xvzcf/tls-interop-runner
[1]: https://boringssl-review.googlesource.com/c/boringssl/+/42984
[2]: https://boringssl-review.googlesource.com/c/boringssl/+/43925

Bug: 275
Change-Id: I49be35af46d1fd5dd9c62252f07d0bae179381ab